fix: adding scroll horizontal
